i2c:phytium: Combine normal mode into master-slave switching mode
This patch combines normal mode into master-slave switching
mode,the application can use master mode or slave mode or
master-slave switching mode.

Mainline: Open-Source
