+++ drivers/i2c/busses/Makefile
@@ -96,7 +96,7 @@ obj-$(CONFIG_I2C_PASEMI)	+= i2c-pasemi-core.o i2c-pasemi-pci.o
 obj-$(CONFIG_I2C_APPLE)		+= i2c-pasemi-core.o i2c-pasemi-platform.o
 obj-$(CONFIG_I2C_PCA_PLATFORM)	+= i2c-pca-platform.o
 obj-$(CONFIG_I2C_PHYTIUM_CORE)	+= i2c-phytium-core.o
-i2c-phytium-core-objs := i2c-phytium-common.o i2c-phytium-master.o i2c-phytium-slave.o
+i2c-phytium-core-objs := i2c-phytium-common.o i2c-phytium-mix.o
 obj-$(CONFIG_I2C_PHYTIUM_PCI)	+= i2c-phytium-pci.o
 obj-$(CONFIG_I2C_PHYTIUM_PLATFORM)	+= i2c-phytium-platform.o
 obj-$(CONFIG_I2C_PNX)		+= i2c-pnx.o
